Assets

+++ Assets/Scripts/TreeCreate.cs
@@ -44,7 +44,7 @@ public class TreeCreate : MonoBehaviour
     public void SetTreeToPlayer()
     {
         Vector3 playerPos = this.transform.position;
-        Vector3Int treePos = new Vector3Int(Mathf.FloorToInt(playerPos.x), (int) Mathf.FloorToInt(playerPos.y), (int) Mathf.FloorToInt(playerPos.z));
+        Vector3Int treePos = new Vector3Int(Mathf.FloorToInt(playerPos.x), Mathf.FloorToInt(playerPos.y), (int) Mathf.FloorToInt(playerPos.z));
         Vector3 cellPos = grid.GetCellCenterWorld(treePos);
         GameObject NewTree = Instantiate(tree, cellPos, Quaternion.identity);
     }
